+++ CHANGELOG.md
@@ -10,6 +10,7 @@ $ npm install @ngxs/store@dev
 - Feature(store): Support async/await syntax for dispatch [#2399](https://github.com/ngxs/store/pull/2399)
 - Feature(store): Improve createSelector signature type checking [#2402](https://github.com/ngxs/store/pull/2402)
 - Feature(store): Export AsyncReturnType from public API [#2405](https://github.com/ngxs/store/pull/2405)
+- Feature(store): Add `safePatch` operator [#2408](https://github.com/ngxs/store/pull/2408)
 - Fix(store): Cleanup observers once subjects complete [#2401](https://github.com/ngxs/store/pull/2401)
 - Fix(store): Skip state mutations when injector is destroyed mid-action [#2406](https://github.com/ngxs/store/pull/2406)
 - Fix(store): Warn when state is mutated after injector destruction [#2407](https://github.com/ngxs/store/pull/2407)